+++ controller/virtualenv.sh
@@ -1,5 +1,29 @@
 #!/bin/bash
 
+# Script to run a single Python command using a specific virtual environment.
+#
+# Copyright 2017-2020 ICTU
+# Copyright 2017-2022 Leiden University
+# Copyright 2017-2024 Leon Helwerda
+# 
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+# 
+#     http://www.apache.org/licenses/LICENSE-2.0
+# 
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+if [ -z $1 ]; then
+	echo "Usage: virtualenv.sh </path/to/venv> [Python command arguments]" >&2
+	echo "Run a single Python command using a specific virtual environment" >&2
+	exit 1
+fi
+
 path=$1
 shift